requirements.txt
streamlit
tensorflow==2.20.0
numpy
pillow
matplotlib
scikit-learn